requirements.txt
fastapi
uvicorn[standard]
python-multipart
jinja2
python-dotenv
openai